also move and rename ResultSet
