+++ sailor/assetcentral/location.py
@@ -4,7 +4,7 @@ Location module can be used to retrieve Location information from AssetCentral.
 Classes are provided for individual Locations as well as groups of Locations (LocationSet).
 """
 
-from .utils import (AssetcentralEntity, _AssetcentralField, ResultSet, _parse_filter_parameters,
+from .utils import (AssetcentralEntity, _AssetcentralField, AssetcentralEntityCollection, _parse_filter_parameters,
                     _fetch_data, _ac_application_url)
 from .._base.masterdata import add_properties
 from ..utils.timestamps import _string_to_timestamp_parser
@@ -38,7 +38,7 @@ class Location(AssetcentralEntity):
     _field_map = {field.our_name: field for field in _LOCATION_FIELDS}
 
 
-class LocationSet(ResultSet):
+class LocationSet(AssetcentralEntityCollection):
     """Class representing a group of Locations."""
 
     _element_type = Location
