+++ sailor/pai/alert.py
@@ -6,8 +6,9 @@ Classes are provided for individual Alert as well as groups of Alerts (AlertSet)
 
 
 from .constants import ALERTS_READ_PATH
-from .utils import PredictiveAssetInsightsEntity, _PredictiveAssetInsightsField, _pai_application_url
-from ..assetcentral.utils import (_fetch_data, _parse_filter_parameters, ResultSet)
+from .utils import (PredictiveAssetInsightsEntity, _PredictiveAssetInsightsField,
+                    PredictiveAssetInsightsEntityCollection, _pai_application_url)
+from ..assetcentral.utils import (_fetch_data, _parse_filter_parameters)
 from .._base.masterdata import add_properties
 from ..utils.timestamps import _odata_to_timestamp_parser
 
@@ -72,7 +73,7 @@ class Alert(PredictiveAssetInsightsEntity):
     _field_map = {field.our_name: field for field in _ALERT_FIELDS}
 
 
-class AlertSet(ResultSet):
+class AlertSet(PredictiveAssetInsightsEntityCollection):
     """Class representing a group of Alerts."""
 
     _element_type = Alert
