+++ sailor/utils/utils.py
@@ -1,5 +1,7 @@
 """Other utility functions that don't fit into any of the specific modules."""
 
+from collections.abc import Iterable
+
 
 # this warning concerns the interactive use case
 class DataNotFoundWarning(Warning):
@@ -8,3 +10,9 @@ class DataNotFoundWarning(Warning):
     def __init__(self, message=None):
         message = "No data found for given parameters." if message is None else message
         super().__init__(message)
+
+
+def _is_non_string_iterable(obj):
+    if issubclass(obj.__class__, str):
+        return False
+    return isinstance(obj, Iterable)
